fix: dereference of `nullptr` due to contested DB
`ScrapeCommand.dbPath` is `null` if no dbPath was explicitly provided as an argument.
Instead `ScrapeCommand.input` is created with an inferred `dbPath`.
This inferred path is never written back to `ScrapeCommand.dbPath`.
When writing the command output, `*ScrapeCommand.dbPath` is serialized to json,
which in most cases will be `nullptr` -> `""`.
This commit makes uses the effective dbpath on `ScrapeCommand.input` instead
to print the correct path.